Add most rubocop rules to ERB files
This way the Ruby code in views will follow our conventions as well.

There are some rules which do not apply to ERB files and so only apply
to .rb files:

* Bundler, Capybara, FactoryBot and RSpec rules do not apply to ERB
  files
* Most Rails rules are meant for models, migrations and controllers; I'm
  only sharing the ones which are generic enough to make sense in ERB
  files
* Rails/OutputSafety is redundant since its functionality is already
  covered by ERB Lint ErbSafety linter
* Layout/LineLength only takes Ruby code into account, so it doesn't really
  measure the actual line length
* Layout/TrailingWhitespace and Layout/TrailingEmptyLines don't work
  properly and their funcionality is covered by ERB Lint
  TralingWhiteSpace, FinalNewline and ExtraNewline linters
* Layout/LeadingEmptyLines and Lint/UselssAssignment generate false
  positives

Even if we don't use any RSpec cops in the `.rubocop_ruby_and_erb.yml`
file, we're still requiring `rubocop-rspec` there because Hound seems to
get confused when there's a `require` statement in both files.

Note that due to a bug in erb_lint 0.0.29 we need to enable all rails
cops, so we're disabling the ones we don't use. We'll remove this code
when Hound uses a newer version of erb_lint.